+++ mopidy/core/tracklist.py
@@ -499,19 +499,19 @@ class TracklistController(object):
         """
         return self._tl_tracks[start:end]
 
-    def mark_playing(self, tl_track):
-        """Method for :class:`mopidy.core.PlaybackController`. **INTERNAL**"""
+    def _mark_playing(self, tl_track):
+        """Internal method for :class:`mopidy.core.PlaybackController`."""
         if self.get_random() and tl_track in self._shuffled:
             self._shuffled.remove(tl_track)
 
-    def mark_unplayable(self, tl_track):
-        """Method for :class:`mopidy.core.PlaybackController`. **INTERNAL**"""
+    def _mark_unplayable(self, tl_track):
+        """Internal method for :class:`mopidy.core.PlaybackController`."""
         logger.warning('Track is not playable: %s', tl_track.track.uri)
         if self.get_random() and tl_track in self._shuffled:
             self._shuffled.remove(tl_track)
 
-    def mark_played(self, tl_track):
-        """Method for :class:`mopidy.core.PlaybackController`. **INTERNAL**"""
+    def _mark_played(self, tl_track):
+        """Internal method for :class:`mopidy.core.PlaybackController`."""
         if self.consume and tl_track is not None:
             self.remove(tlid=[tl_track.tlid])
             return True